refactor: migrate logger from winston to pino
