+++ package.json
@@ -9,7 +9,7 @@
     "start": "tsnd --respawn --exit-child -r dotenv/config -r tsconfig-paths/register src/web/api/server.ts",
     "test": "jest",
     "test:unit": "jest --selectProjects unit-tests",
-    "test:integration": "jest --selectProjects integration-tests",
+    "test:functional": "jest --selectProjects functional-tests",
     "prepare": "husky install"
   },
   "dependencies": {
@@ -17,8 +17,8 @@
     "awilix": "^8.0.1",
     "express": "^4.18.2",
     "helmet": "^7.1.0",
+    "pino": "^8.18.0",
     "swagger-ui-express": "^5.0.0",
-    "winston": "^3.11.0",
     "yup": "^1.3.3"
   },
   "devDependencies": {
