--- src/application/common/interfaces/ILogger.ts
@@ -1,14 +1,0 @@
-type LogMethod = {
-  (message: { [key: string]: unknown }): void;
-};
-
-export interface ILogger {
-  alert: LogMethod;
-  crit: LogMethod;
-  debug: LogMethod;
-  emerg: LogMethod;
-  error: LogMethod;
-  info: LogMethod;
-  notice: LogMethod;
-  warning: LogMethod;
-}
