+++ src/application/common/interfaces/posts-repository.ts
@@ -1,6 +1,6 @@
 import { Post } from '@domain/entities';
 
-export interface IPostsRepository {
+export interface PostsRepository {
   create(parameters: { post: Post }): Promise<{ id: number }>;
   delete(parameters: { id: number }): Promise<void>;
   getById(parameters: { id: number }): Promise<Post | null>;
