+++ src/application/posts/queries/get-post/__tests__/get-post-query.test.ts
@@ -1,12 +1,12 @@
 import { mockDeep } from 'jest-mock-extended';
-import { IPostsRepository } from '@application/common/interfaces';
+import { PostsRepository } from '@application/common/interfaces';
 import { NotFoundException } from '@application/common/exceptions';
 import { Post } from '@domain/entities';
 import { makeGetPostQuery } from '../get-post-query';
 
 describe('getPostQuery', () => {
   function setup() {
-    const postsRepository = mockDeep<IPostsRepository>({
+    const postsRepository = mockDeep<PostsRepository>({
       getById: jest.fn().mockResolvedValue(
         new Post({
           id: 1,
